\name{Felix-package}
\alias{Felix-package}
\alias{Felix}
\docType{package}
\title{
\packageTitle{Felix}
}
\description{
\packageDescription{Felix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Felix}
\packageIndices{Felix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Felix}

Maintainer: \packageMaintainer{Felix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
